Change and split the Vacancies page and add the VacancyCard Page
