+++ FrontEnd/src/components/hr/Vacancies.jsx
@@ -1,19 +1,11 @@
+// src/pages/Vacancies.jsx
 import React, { useState } from "react";
-import {
-  Plus,
-  Search,
-  Filter,
-  Edit,
-  Trash2,
-  MapPin,
-  Calendar,
-  Users,
-  DollarSign,
-  Eye,
-} from "lucide-react";
+import { Plus, Search } from "lucide-react";
 import { mockVacancies } from "../../data/mockData";
+import VacancyCard from "./VacancyCard";
 
- function Vacancies() {
+
+function Vacancies() {
   const [vacancies, setVacancies] = useState(mockVacancies);
   const [searchTerm, setSearchTerm] = useState("");
   const [filterStatus, setFilterStatus] = useState("All");
@@ -21,8 +13,8 @@ import { mockVacancies } from "../../data/mockData";
   const [selectedVacancy, setSelectedVacancy] = useState(null);
 
   const filteredVacancies = vacancies.filter((vacancy) => {
-    const title = vacancy.title ? vacancy.title.toLowerCase() : '';
-    const department = vacancy.department ? vacancy.department.toLowerCase() : '';
+    const title = vacancy.title ? vacancy.title.toLowerCase() : "";
+    const department = vacancy.department ? vacancy.department.toLowerCase() : "";
     const matchesSearch =
       title.includes(searchTerm.toLowerCase()) ||
       department.includes(searchTerm.toLowerCase());
@@ -33,6 +25,7 @@ import { mockVacancies } from "../../data/mockData";
 
   return (
     <div className="px-4 sm:px-6 lg:px-8">
+      {/* Header */}
       <div className="sm:flex sm:items-center sm:justify-between">
         <div>
           <h1 className="text-2xl font-bold text-gray-900">Vacancies</h1>
@@ -89,490 +82,6 @@ import { mockVacancies } from "../../data/mockData";
             }
           />
         ))}
-      </div>
-
-      {showCreateModal && (
-        <CreateVacancyModal
-          onClose={() => setShowCreateModal(false)}
-          onSave={(vacancy) => {
-            setVacancies([
-              ...vacancies,
-              { ...vacancy, id: Date.now().toString() },
-            ]);
-            setShowCreateModal(false);
-          }}
-        />
-      )}
-
-      {selectedVacancy && (
-        <EditVacancyModal
-          vacancy={selectedVacancy}
-          onClose={() => setSelectedVacancy(null)}
-          onSave={(updatedVacancy) => {
-            setVacancies(
-              vacancies.map((v) =>
-                v.id === updatedVacancy.id ? updatedVacancy : v
-              )
-            );
-            setSelectedVacancy(null);
-          }}
-        />
-      )}
-    </div>
-  );
-}
-
-
-
-
-function VacancyCard({ vacancy, onEdit, onDelete }) {
-  const getStatusColor = (status) => {
-    switch (status) {
-      case "Open":
-        return "bg-green-100 text-green-800";
-      case "Closed":
-        return "bg-red-100 text-red-800";
-      case "On Hold":
-        return "bg-yellow-100 text-yellow-800";
-      default:
-        return "bg-gray-100 text-gray-800";
-    }
-  };
-
-  return (
-    <div className="bg-white rounded-lg shadow hover:shadow-md transition-shadow duration-200">
-      <div className="p-6">
-        <div className="flex items-center justify-between mb-4">
-          <span
-            className={`inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium ${getStatusColor(
-              vacancy.status
-            )}`}
-          >
-            {vacancy.status}
-          </span>
-          <div className="flex space-x-2">
-            <button
-              onClick={() => onEdit(vacancy)}
-              className="text-gray-400 hover:text-gray-600"
-            >
-              <Edit className="h-4 w-4" />
-            </button>
-            <button
-              onClick={() => onDelete(vacancy.id)}
-              className="text-gray-400 hover:text-red-600"
-            >
-              <Trash2 className="h-4 w-4" />
-            </button>
-          </div>
-        </div>
-
-        <h3 className="text-lg font-semibold text-gray-900 mb-2">
-          {vacancy.title}
-        </h3>
-        <p className="text-sm text-gray-600 mb-4">{vacancy.department}</p>
-
-        <div className="space-y-2 mb-4">
-          <div className="flex items-center text-sm text-gray-500">
-            <MapPin className="h-4 w-4 mr-2" />
-            {vacancy.location}
-          </div>
-          <div className="flex items-center text-sm text-gray-500">
-            <Calendar className="h-4 w-4 mr-2" />
-            Deadline: {vacancy.deadline ? new Date(vacancy.deadline).toLocaleDateString() : 'N/A'}
-          </div>
-          <div className="flex items-center text-sm text-gray-500">
-            <DollarSign className="h-4 w-4 mr-2" />$
-            {typeof vacancy.salary === 'number' ? vacancy.salary.toLocaleString() : ''}
-          </div>
-        </div>
-
-        <div className="flex items-center justify-between pt-4 border-t border-gray-200">
-          <div className="flex items-center text-sm text-gray-500">
-            <Users className="h-4 w-4 mr-1" />
-            {vacancy.applicantCount} applicants
-          </div>
-          <button className="inline-flex items-center text-sm text-blue-600 hover:text-blue-500">
-            <Eye className="h-4 w-4 mr-1" />
-            View Details
-          </button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function CreateVacancyModal({ onClose, onSave }) {
-  const [formData, setFormData] = useState({
-    title: "",
-    department: "",
-    location: "",
-    type: "Full-time",
-    description: "",
-    responsibilities: "",
-    salary: "",
-    deadline: "",
-  });
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    const vacancy = {
-      title: formData.title,
-      department: formData.department,
-      location: formData.location,
-      type: formData.type,
-      description: formData.description,
-      responsibilities: formData.responsibilities
-        .split("\n")
-        .filter((r) => r.trim()),
-      salary: parseInt(formData.salary),
-      status: "Open",
-      postedDate: new Date().toISOString().split("T")[0],
-      deadline: formData.deadline,
-      applicantCount: 0,
-      createdBy: "HR001", // You can modify this as needed
-    };
-    onSave(vacancy);
-  };
-
-  return (
-    <div className="fixed inset-0 bg-gray-600 bg-opacity-50 overflow-y-auto h-full w-full z-50">
-      <div className="relative top-20 mx-auto p-5 border w-full max-w-2xl shadow-lg rounded-md bg-white">
-        <div className="mt-3">
-          <h3 className="text-lg font-medium text-gray-900 mb-4">
-            Create New Vacancy
-          </h3>
-          <form onSubmit={handleSubmit} className="space-y-4">
-            <div className="grid grid-cols-1 gap-4 sm:grid-cols-2">
-              <div>
-                <label className="block text-sm font-medium text-gray-700">
-                  Job Title
-                </label>
-                <input
-                  type="text"
-                  required
-                  value={formData.title}
-                  onChange={(e) =>
-                    setFormData({ ...formData, title: e.target.value })
-                  }
-                  className="mt-1 block w-full border border-gray-300 rounded-md px-3 py-2 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                />
-              </div>
-              <div>
-                <label className="block text-sm font-medium text-gray-700">
-                  Department
-                </label>
-                <input
-                  type="text"
-                  required
-                  value={formData.department}
-                  onChange={(e) =>
-                    setFormData({ ...formData, department: e.target.value })
-                  }
-                  className="mt-1 block w-full border border-gray-300 rounded-md px-3 py-2 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                />
-              </div>
-            </div>
-
-            <div className="grid grid-cols-1 gap-4 sm:grid-cols-2">
-              <div>
-                <label className="block text-sm font-medium text-gray-700">
-                  Location
-                </label>
-                <input
-                  type="text"
-                  required
-                  value={formData.location}
-                  onChange={(e) =>
-                    setFormData({ ...formData, location: e.target.value })
-                  }
-                  className="mt-1 block w-full border border-gray-300 rounded-md px-3 py-2 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                />
-              </div>
-              <div>
-                <label className="block text-sm font-medium text-gray-700">
-                  Employment Type
-                </label>
-                <select
-                  value={formData.type}
-                  onChange={(e) =>
-                    setFormData({ ...formData, type: e.target.value })
-                  }
-                  className="mt-1 block w-full border border-gray-300 rounded-md px-3 py-2 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                >
-                  <option value="Full-time">Full-time</option>
-                  <option value="Part-time">Part-time</option>
-                  <option value="Contract">Contract</option>
-                  <option value="Internship">Internship</option>
-                </select>
-              </div>
-            </div>
-
-            <div>
-              <label className="block text-sm font-medium text-gray-700">
-                Job Description
-              </label>
-              <textarea
-                required
-                rows={3}
-                value={formData.description}
-                onChange={(e) =>
-                  setFormData({ ...formData, description: e.target.value })
-                }
-                className="mt-1 block w-full border border-gray-300 rounded-md px-3 py-2 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-              />
-            </div>
-
-            <div>
-              <label className="block text-sm font-medium text-gray-700">
-                Responsibilities (one per line)
-              </label>
-              <textarea
-                required
-                rows={3}
-                value={formData.responsibilities}
-                onChange={(e) =>
-                  setFormData({ ...formData, responsibilities: e.target.value })
-                }
-                className="mt-1 block w-full border border-gray-300 rounded-md px-3 py-2 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                placeholder="Manage team projects\nDevelop new features\nCollaborate with stakeholders"
-              />
-            </div>
-
-            <div className="grid grid-cols-1 gap-4 sm:grid-cols-2">
-              <div>
-                <label className="block text-sm font-medium text-gray-700">
-                  Offered Salary ($)
-                </label>
-                <input
-                  type="number"
-                  required
-                  value={formData.salary}
-                  onChange={(e) =>
-                    setFormData({ ...formData, salary: e.target.value })
-                  }
-                  className="mt-1 block w-full border border-gray-300 rounded-md px-3 py-2 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                />
-              </div>
-              <div>
-                <label className="block text-sm font-medium text-gray-700">
-                  Application Deadline
-                </label>
-                <input
-                  type="date"
-                  required
-                  value={formData.deadline}
-                  onChange={(e) =>
-                    setFormData({ ...formData, deadline: e.target.value })
-                  }
-                  className="mt-1 block w-full border border-gray-300 rounded-md px-3 py-2 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                />
-              </div>
-            </div>
-
-            <div className="flex justify-end space-x-3 pt-4">
-              <button
-                type="button"
-                onClick={onClose}
-                className="px-4 py-2 border border-gray-300 rounded-md text-sm font-medium text-gray-700 bg-white hover:bg-gray-50"
-              >
-                Cancel
-              </button>
-              <button
-                type="submit"
-                className="px-4 py-2 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-blue-600 hover:bg-blue-700"
-              >
-                Create Vacancy
-              </button>
-            </div>
-          </form>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function EditVacancyModal({ vacancy, onClose, onSave }) {
-  const [formData, setFormData] = useState({
-    title: vacancy.title,
-    department: vacancy.department,
-    location: vacancy.location,
-    type: vacancy.type,
-    description: vacancy.description,
-    responsibilities: vacancy.responsibilities.join("\n"),
-    salary: vacancy.salary.toString(),
-    deadline: vacancy.deadline,
-    status: vacancy.status,
-  });
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    const updatedVacancy = {
-      ...vacancy,
-      title: formData.title,
-      department: formData.department,
-      location: formData.location,
-      type: formData.type,
-      description: formData.description,
-      responsibilities: formData.responsibilities
-        .split("\n")
-        .filter((r) => r.trim()),
-      salary: parseInt(formData.salary),
-      deadline: formData.deadline,
-      status: formData.status,
-    };
-    onSave(updatedVacancy);
-  };
-
-  return (
-    <div className="fixed inset-0 bg-gray-600 bg-opacity-50 overflow-y-auto h-full w-full z-50">
-      <div className="relative top-20 mx-auto p-5 border w-full max-w-2xl shadow-lg rounded-md bg-white">
-        <div className="mt-3">
-          <h3 className="text-lg font-medium text-gray-900 mb-4">
-            Edit Vacancy
-          </h3>
-          <form onSubmit={handleSubmit} className="space-y-4">
-            <div className="grid grid-cols-1 gap-4 sm:grid-cols-2">
-              <div>
-                <label className="block text-sm font-medium text-gray-700">
-                  Job Title
-                </label>
-                <input
-                  type="text"
-                  required
-                  value={formData.title}
-                  onChange={(e) =>
-                    setFormData({ ...formData, title: e.target.value })
-                  }
-                  className="mt-1 block w-full border border-gray-300 rounded-md px-3 py-2 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                />
-              </div>
-              <div>
-                <label className="block text-sm font-medium text-gray-700">
-                  Status
-                </label>
-                <select
-                  value={formData.status}
-                  onChange={(e) =>
-                    setFormData({ ...formData, status: e.target.value })
-                  }
-                  className="mt-1 block w-full border border-gray-300 rounded-md px-3 py-2 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                >
-                  <option value="Open">Open</option>
-                  <option value="Closed">Closed</option>
-                  <option value="On Hold">On Hold</option>
-                </select>
-              </div>
-            </div>
-
-            <div className="grid grid-cols-1 gap-4 sm:grid-cols-2">
-              <div>
-                <label className="block text-sm font-medium text-gray-700">
-                  Department
-                </label>
-                <input
-                  type="text"
-                  required
-                  value={formData.department}
-                  onChange={(e) =>
-                    setFormData({ ...formData, department: e.target.value })
-                  }
-                  className="mt-1 block w-full border border-gray-300 rounded-md px-3 py-2 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                />
-              </div>
-              <div>
-                <label className="block text-sm font-medium text-gray-700">
-                  Location
-                </label>
-                <input
-                  type="text"
-                  required
-                  value={formData.location}
-                  onChange={(e) =>
-                    setFormData({ ...formData, location: e.target.value })
-                  }
-                  className="mt-1 block w-full border border-gray-300 rounded-md px-3 py-2 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                />
-              </div>
-            </div>
-
-            <div>
-              <label className="block text-sm font-medium text-gray-700">
-                Job Description
-              </label>
-              <textarea
-                required
-                rows={3}
-                value={formData.description}
-                onChange={(e) =>
-                  setFormData({ ...formData, description: e.target.value })
-                }
-                className="mt-1 block w-full border border-gray-300 rounded-md px-3 py-2 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-              />
-            </div>
-
-            <div>
-              <label className="block text-sm font-medium text-gray-700">
-                Responsibilities (one per line)
-              </label>
-              <textarea
-                required
-                rows={3}
-                value={formData.responsibilities}
-                onChange={(e) =>
-                  setFormData({ ...formData, responsibilities: e.target.value })
-                }
-                className="mt-1 block w-full border border-gray-300 rounded-md px-3 py-2 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-              />
-            </div>
-
-            <div className="grid grid-cols-1 gap-4 sm:grid-cols-2">
-              <div>
-                <label className="block text-sm font-medium text-gray-700">
-                  Offered Salary ($)
-                </label>
-                <input
-                  type="number"
-                  required
-                  value={formData.salary}
-                  onChange={(e) =>
-                    setFormData({ ...formData, salary: e.target.value })
-                  }
-                  className="mt-1 block w-full border border-gray-300 rounded-md px-3 py-2 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                />
-              </div>
-              <div>
-                <label className="block text-sm font-medium text-gray-700">
-                  Application Deadline
-                </label>
-                <input
-                  type="date"
-                  required
-                  value={formData.deadline}
-                  onChange={(e) =>
-                    setFormData({ ...formData, deadline: e.target.value })
-                  }
-                  className="mt-1 block w-full border border-gray-300 rounded-md px-3 py-2 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                />
-              </div>
-            </div>
-
-            <div className="flex justify-end space-x-3 pt-4">
-              <button
-                type="button"
-                onClick={onClose}
-                className="px-4 py-2 border border-gray-300 rounded-md text-sm font-medium text-gray-700 bg-white hover:bg-gray-50"
-              >
-                Cancel
-              </button>
-              <button
-                type="submit"
-                className="px-4 py-2 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-blue-600 hover:bg-blue-700"
-              >
-                Save Changes
-              </button>
-            </div>
-          </form>
-        </div>
       </div>
     </div>
   );
